Update AdaptiveRotationHyperConnection to ensure proper dtype handling in normalization step

+++ records/track_10min_16mb/2026-04-09_SP8192_3LayerRecur_ParResid_QK525_LegalTTT/train_gpt_expanded.py
@@ -528,8 +528,8 @@ class AdaptiveRotationHyperConnection(nn.Module):
         B, T, n, D = streams.shape
 
         # LayerNorm on the full flattened stream state (Eq 35)
-        xflat = streams.flatten(-2, -1).float()   # [B, T, n*D]
-        normed = self.norm(xflat)                 # [B, T, n*D]
+        xflat = streams.flatten(-2, -1).float()        # [B, T, n*D]
+        normed = self.norm(xflat).to(streams.dtype)    # [B, T, n*D], cast back for weight matmuls
 
         # H_pre: row-stochastic n×n — aggregates streams into one input (Eq 36, 39)
         h_pre = F.softmax(
